regkas-server/regkas-server-domain/regkas-server-domain-core/src/main/java/org/regkas/domain/core/model/ReceiptTypeSammelCore.java
package org.regkas.domain.core.model;

import com.google.common.collect.Lists;
import org.regkas.domain.api.receipt.precondition.Precondition;
import org.regkas.domain.api.receipt.precondition.SignatureDeviceAvailable;
import org.regkas.domain.api.receipt.precondition.StartReceiptAvailable;
import org.regkas.model.ReceiptTypeSammelEntity;
import org.regkas.domain.api.model.ReceiptTypeSammel;
import org.regkas.domain.api.turnovercounter.EncryptTurnoverCounter;
import org.regkas.domain.api.turnovercounter.UpdateTurnoverCounter;
import org.regkas.domain.core.turnovercounter.EncryptTurnoverCounterDefault;
import org.regkas.domain.core.turnovercounter.UpdateTurnoverCounterNothing;

import java.util.List;

public class ReceiptTypeSammelCore extends ReceiptTypeCore<ReceiptTypeSammel, ReceiptTypeSammelEntity> implements ReceiptTypeSammel {

    public ReceiptTypeSammelCore(ReceiptTypeSammelEntity receiptType) {
        super(receiptType);
    }

    @Override
    public List<Class<? extends Precondition>> getPreconditions() {
        return Lists.newArrayList(StartReceiptAvailable.class, SignatureDeviceAvailable.class);
    }

    @Override
    public UpdateTurnoverCounter getUpdateTurnoverCounter() {
        return fromCDI(UpdateTurnoverCounterNothing.class);
    }

    @Override
    public EncryptTurnoverCounter getEncryptTurnoverCounter() {
        return fromCDI(EncryptTurnoverCounterDefault.class);
    }
}
